Large number of changes, see commit details
The configuration files have been changed to use a Hikari JBDC connection pooling.

Twin data sources implemented, entityManagers can write to the local file based db or to a database server. There is the primary data-source which is defaulted for Hibernate to rely on, but for any EntityInterfaces/Repositories you can specify a different EntityManager and use the secondary data-source. The main rason for htis was for easier data-source switching for standalone production releases. In this commit all the Secondary datasource is disabled by commenting out the code. The ability to quickly switch the primary data-source remains. Documented in readme. It has been tested and does work.

Structure of Menu classes changed where now method chains require User from loggedInMenu as far as until an Avatar is selected, and then an Avatar onwards from there. There are no stored pointers to entities in the Service or Menu (Management) classes.

User's list of Avatars is EAGER loaded instead of LAZY due to session issues. See devnote #4.

The dungeons are free of any particular Avatar or User. This means that when an avatar is assigned a currDungeon there has to be a selection process for that. Currently that process is

Debug mode is now read from application.properties and allows setting for speeding up and simplifying event sequences in game testing.

Added new file application_example.properties to repository for easy distribution.

Finally removed un-needed files from repository and optimized imports.